requirement.txt
backcall==0.1.0
decorator==4.3.0
Django==2.0
html5lib==1.0b8
httplib2==0.11.3
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.0
pdfkit==0.6.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
#pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pudb==2017.1.4
Pygments==2.2.0
PyPDF2==1.26.0
python-dateutil==2.7.2
pytz==2018.4
reportlab==3.4.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urwid==2.0.1
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.2.2
